chore(build): Add support for RPM-based distros #1335
This adds support for building the RPM.
Apache service is named httpd on Red Hat and apache2 on Ubuntu. This
change makes it possible to deal with both service names. Since we
don't have the a2ensite/a2enmod on Red Hat, we have to manually create
the logic to enable the spinnaker.conf site.